src/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
requests==2.23.0
robotframework==4.1.2
robotframework-requests==0.7.0
urllib3==1.25.9
robotframework-seleniumlibrary==5.1.3
robotframework-xvfb==1.2.2
